(progress) Can load tailwind colors into picker
The get_color_under_cursor function can now accurately determine the tailwind color under the cursor.
Super tricky because of all the tailwind specific prefixes.
The even harder part is matching in a way that let's me get just the color/shade from the color.
I also changed the data structure to get the prefix name, start, and end.
This may come in handy later to maybe.
It would be cool to get the telescope extension change the current prefix under the cursor as well as the color.